+++ TVRename/Custom Controls/NativeMethods.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace TVRename
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport("user32.dll")]
+        internal static extern int GetScrollPos(IntPtr hWnd, int nBar);
+
+        [DllImport("user32.dll")]
+        internal static extern IntPtr SendMessage(IntPtr hWnd, uint msg, IntPtr wParam, IntPtr lParam);
+
+        // MAH: Added in support of the Filter TextBox Button
+        [DllImport("user32.dll")]
+        internal static extern IntPtr SendMessage(IntPtr hWnd, int msg, IntPtr wp, IntPtr lp);
+    }
+}
